Disable Windows github actions caching.
It already didn't work anyway, but now it seems to be actively causing problems by randomly failing due to lack of disk space. Though how it can only sometimes run out of disk space despite trying to cache the same thing every time, and when the runner instance is supposed to always have the same amount of disk space available, is anyone's guess.