font: Fix double resource leak in font::pango_text
The impacted code path is only executed as part of the implementation of
the characters_per_line constraint.

According to the Pango documentation, PangoFontMetrics should be
released using pango_font_metrics_unref(), and PangoFont should be
released using *something* not specified. Existing code, and the
documentation for pango_fontset_get_font(), suggests the answer for the
latter is g_object_unref().