Simplify pango_text: surface is always dirty, remove unused functionality
Every instance of pango_text had its own surface and dirty bit for caching the
result of rendering. However, pango_text is used as a singleton, so as soon as
there are two strings on screen then they're overwriting each others' cached result
and the dirty bit is always set. The loading screen is probably the only screen
where the dirty bit could be useful, as there's only one string on that screen.

The other parts removed were simply unused:
* the `force` options
* insert_unicode
* outlines