pango_text: remove unused force options and single-character insertion
The single-character insertion seems unlikely to be used in future. While
pango_text isn't a singleton, a few shared instances are created and then used
in a singleton-like way, with each caller assuming that it received it in an
unknown state and always calling set_text(), set_font_size(), set_font_style(),
etc.