Fix attack prediction assert
The corrections applied in calculate_probability_of_debuff() are already taken into account in the main attack outcome calculation for being slowed.  They do not need to be applied again, that only needs to be done for being poisoned.  Doing this for being slowed also can cause an assert, such as the one reported in issue #5533 (and probably also the one in issue #5002).

Fixes #5533