PROTOTYPE: Handle hotkeys on non-Latin keyboards via transliteration
Prototype: no UI is added for configuring this; at the moment it comes with
a Python script and you need to paste the output of that script into a .cfg
somewhere - for example data/core/hotkeys.cfg.

Prototype: generated data is included for:
* Greek to Latin
* Russian to Qwerty
* Ukrainian to Qwerty (possibly wrong, Wikipedia and IBus seem to disagree
   about which letters change between the Russian and Ukrainian layouts)

Imagine a keyboard with Latin "a" and Greek "α" on the same key, "b" and "β" on
the same key, etc. This commit adds a fallback so that, when the user presses
ctrl+β, the engine first searches for ctrl+β and then falls back to searching
for ctrl+b.

The transliterations are only searched for when the original input doesn't
trigger a hotkey; if the user manually adds a hotkey for ctrl+β then it will be
found first when the keyboard is in Greek mode.

I'm using Greek for the examples because (looking at Wikipedia) there seems to
be consistency in which Greek letters match to which Latin letters. Cyrillic on
the other hand seems to have as many keyboard variants as Latin does (Qwerty,
Qwertz, Azerty, etc), and the user may expect the Latin letters on their keyboard
to be (at least) Qwerty or Qwertz.

The Greek layout puts a semicolon on the Latin-q key, which has been omitted
from this conversion. Even if it was included, it would look for the semicolon
and colon based hotkeys before looking for the q hotkeys, that's a design
choice to avoid overriding existing hotkeys.

The syntax for the preferences file is like this:

```
[hotkey_transliteration]
	name=_ "Greek letters to Latin letters"
	[substitute]
		src=α
		dst=a
	[/substitute]
	[substitute]
		src=β
		dst=b
	[/substitute]
	[substitute]
		src=ψ
		dst=c
	[/substitute]
	[substitute]
		src=δ
		dst=d
	[/substitute]
	...
[/hotkey_transliteration]
```

Many thanks to Renom for investigating a different method of handling keyboard
layouts. Although that method ended up blocked by SDL issues, the work saved me
from trying to read the actual keyboard layouts via SDL for this method.