Rendering only a viewport area in pango_text
This is meant to be an optimisation, which will reduce memory usage when
rendering text. It doesn't seem to reduce the amount of work done by Pango, but
the drawing will be clipped in the Cairo library.

The line-by-line workaround for rendering huge areas of text is dropped, as we
now only expect to be rendering screen-sized or smaller parts of the text.
Rendering huge texts (greater than 64kpixels high) requires the caller of
pango_text::render() to change to calling pango_text::render(SDL_Rect) instead,
which will be handled in a separate commit - fixing the known crash when
opening the credits screen will require that commit on top of this one.