EI S2: Lift shroud (but not fog) around the amulet
Gives the player a hint to explore around the amulet, despite the objectives
saying "Get moving quickly", which discourages that in the first playthrough.

This might also prevent players interpreting the revealed path directly through
the trolls as the suggested path through the level. By putting one of the dots
in an obvious dead-end, it's less of a surprise that the path shown by the
other dots leads to getting trapped between trolls and undead.

Part of #4145