EI S07b: Wake the sea creatures with noise instead of magic
This comes from a discussion about which archaic-sounding pronoun should be
used instead of "you" in "Creatures from the deep, I summon you to destroy this
foe". The conclusion with that was that "thee" or "ye" would sound better, but
"you" is the correct archaic pronoun.

The .ogg file is copied from Northern Rebirth.