LoW The Chief Must Die: replace Formula AI with Micro AI
Formula AI is not maintained any more and will be removed in the future.  This replaces the Patrol FAI use in this scenario with a Patrol Micro AI.

Note that the FAI guard_radius=3 is equivalent to the MAI attack_range=4 as the former is the distance to the hex from which the patrol attacks, while the latter is the distance to the enemy.