bi: Add Lua library version
It *is* a built-in library, but it changes over time. It may prove
useful in future times to be able to readily tell what Lua version is
attached to a particular Wesnoth dev version since it's harder to keep
track of singular releases (e.g. I don't remember which dev release
upgraded us to Lua 5.4 right now).