Prevent deprecation message details from being on a separate line.
The log handler does some kind of per-line sorting of these, so a single deprecation warning ends up with the detail line first and the message line second (line order reversed). When there are multiple deprecation messages, their detail lines and message lines get jumbled together.  This seems to be tied together with how it detects repeated instances of the exact same error/warning, so putting both the message and the detail on the same line is much simpler than tinkering with that.