village support +1 for 2 player mode (#5)
this sort of makes Dark Forecast passable for 2 player mode. Initially, it was only passable if it was one-player-mode as one player could have 11 villages active and could get a good army built prior to the final wave.
2 player suffered from lack of resources to replace losses and would be unwinnable in some cases.
Only 2 cases where it was won in 2 player mode was when the players were of exceptionally high skill level but average skill-levelled players have yet to even win.