font: Cleanup link formatting logic
Let's internally handle this mess in terms of tokens instead of
delimiter positions.

This alters the code so we don't add and substract ones (1) all over the
place. It also simplifies the final std::string_view::substr() call so
we don't need to count the remainder's length at all.

And more importantly, this makes the case where no delimiters had been
previously found trivial instead of causing it to break and result in no
text being output at all. We also don't start with a size_t value of -1
(actually a very large unsigned value).

Closes #5625.