Remove workarounds for SDL2.0.6 bug in SDL_FreeSurface()
Workarounds in CVideo class and surface::free_surface() method are not
required after SDL2 version bump to 2.0.8.

The workaround in CVideo class was incorrectly applied for SDL2 >= 2.0.6,
instead exactly for 2.0.6, as in the case of surface::free_surface().

After removing workarounds, the surface::clear_without_free() method
has no callers and can be removed.