add an [anti_special] ability, which can negate the effects of certains specials weapons
(slow, poison, drains, plague and petrifies) are the only ones affected.

They are not here to provide protection against all status changes, but rather to be able to cancel certain weapons.

like protect against certain poisons or slow attacks (but not all, whereas with the status unpoisonable, it is all or nothing).