Icemonax advances to Greater Icemonax
I think it is perfectly reasonable to be able to have icemonax advance to greater icemonax. This makes the units feel like a proper unit line, and gives UMC authors more freedom how to use the unit. To compensate for the fact that a lvl0 unit advances to lvl1, I set the exp quite high (53) to make this not something abusable