Move path-related functions into a new wesnoth.paths module
This also adds support to the mapgen version of wesnoth.find_path for
taking an options table as the 3rd argument instead of several additional arguments.

The intent of this is to bring the two versions of find_path closer together.
For now, however, the actual options accepted in the table have not changed.