Create a [anti_special] ability who negate specifically some specials.
With this ability, it is possible to negate the effects of specials without ipmlement filter in this special
and apply that to core specials or abilities used like specials.
Another advantage is the [affect_adjacent] but also to affect a special unique without affect special of same type who could be used in same attack.
an exemple can be to cancel the effect of magical chance_to_hit without affect the marsksman abilities or to create different type of poisons