WML Schema: t_string is now a built-in type and no longer attempts a regex match
A key validates as type t_string if one of the following is true:
- The key is not present
- The key has at least one segment with a translation mark
- The key is blank (an empty string)

Any type other than t_string is not allowed to be translatable,
unless you specify allow_translatable=yes in the [type] tag.

An optionally-translatable string could also be defined
as a union of t_string and some other type.