WFL: Add functions to get the time of day bonus
tod_bonus: accounts for inherent terrain lighting and unit illumination effects
base_tod_bonus: ignores these effects

Both take an optional location and an optional turn number. With a location, time areas are taken into account. With a turn number, you can check what the bonus will be iin the future, or what it was in the past.

Closes #5776