Add-ons manager: Fix bugs that return to the title screen with an "Invalid WML" message
This fixes two specific causes of #3059, where code assumes that all parts of
the add-ons manager's UI are accessible, an assumption that fails when the
dialog uses a stacked widget to handle small window sizes.

It would be better to redesign the C++ for this dialog, but a quick workaround
for the observed issues is much easier.

* Fixes #5810, triggered by using the version drop-down in small-screen mode.
* Adds a workaround for keyboard input changing the selected add-on on small screens.