When starting or loading a game, warn if the next scenario is unknown
Check that the scenario's next_scenario= exists, and display a warning
if it would lead to an "Unknown Scenario" error. This also checks any
[endlevel] tags in events.

The error message includes the id of the missing scenario, but is mainly
a recommendation to check the forums and report a bug in the campaign.

This version has a short, one-size-fits-all, text for the error message.