Slight optimisation of canvas::set_variables using std::move
Doesn't make much difference for release builds built with GCC,
it takes between 1% and 7% off the cost of std::update_canvas().
Calculated using the time taken by gui2::widget::get_best_size()
as a reference.

It's a tiny improvement, but it's also a tiny change, and the
improvement is in the loop that's MP lobby's performance problem,
as described in issue #5578.