S05: Stop prey taking drake villages
The wildlife AI does a good job of avoiding moving to villages, but very occasionally it happens somehow (possibly the spawner dumps them there, which the AI has no control over). When it does happen, capture the village for the player to prevent wildlife from stealing the player's villages in the worst case scenario. Giving the village to $owner_side would be a better option except it causes weird bugs