gui2/addon_manager: Add TLS status and server address indicator
This is displayed on the bottom left in "big GUI" mode and on the bottom
middle in "small GUI" mode. The hostname label's colour depends on
whether the connection is actually layered in TLS or not.

The tooltips are kind of a hack because we currently can't set a tooltip
on a grid so it triggers whenever the player mouses over any of its
children or the border/margin space between cells. Instead we set the
same tooltips for both widget pairs via preprocessor trickery and call
it a day.