Title screen: move "change language" back to the text-button area
Okay for the string-freeze - the necessary string was re-added in fc50458b.

No forward-port required, as master will have a full title screen redesign in
issue #4531. This commit is reverting acc96b16, which is a part of #4531,
because the redesign has paused at a place that causes problems with large
screens - the button added in acc96b16 moves further from the center of the
screen as the screen size increases, and it doesn't scale with screen size. On
an 800x600 window it's easily noticeable, but on 2560x1440 it's less obvious
than either Dan Tonk or Weldyn.