SotA: Synchronise the zombie recruitment dialog
This dialog works by setting variables in the postshow function,
so the code had to be moved around to ensure that those variables
were synchronised, not just the return value of the dialog.

The message about not having enough gold isn't necessary for the
replay, so that gets moved into the synchronised code too.