AI: Support [modify_side][ai][modify_ai]/[micro_ai]
Although there is little reason to do this in practice, there is merit in supporting all the same syntax in both places.

For example, if AI setup is defined as a macro, it could contain [modify_ai] tags.
This change means that placing that macro in [modify_side][ai] has the same effect as placing it in [side][ai].