Clarify objectives (resolves #5608)
According to the bonus objective as it stands, it should be possible to end the scenario early by simply destroying the 'generator' stones even if leaving a few skeletons around.
However, the way victory condition is established simply by counting the number of enemy units at the start of a turn - if there are no units left, then the stones are assumed to be destroyed. (No generator stones active -> no enemy units)
As a work-around, simply update the objectives to make it clear that all enemies need to be eliminated for victory to be achieved.